fix deleting of image after editing
I just made a boolean check to only alter database when there actually has been a change to photo

+++ src/screens/ArtisanHub/actions.js
@@ -69,7 +69,7 @@ export function fetchArtisans(cgaID) {
 
 //Updates artisan info and if image is passed in than delete current one in storage,
 //update it with new image picked in both storage and database
-export const saveArtisan = ({ name, phoneNumber, location, description, profilePicturePath, uid }) => {
+export const saveArtisan = ({ name, phoneNumber, location, description, profilePicturePath, uid, changedImage }) => {
   return (dispatch) => {
     return new Promise(async (resolve) => {
       await firebase.database().ref(`/artisans/${uid}`)
@@ -84,13 +84,12 @@ export const saveArtisan = ({ name, phoneNumber, location, description, profileP
       }
       
       //UPDATE PHONE MAPPING HERE OR IN FUNCTIONS
-
-      if (profilePicturePath) {
+      if (profilePicturePath && changedImage) {
         await firebase.storage().ref(`artisanFiles/${uid}/images/profilePicture`).delete().catch(error => console.log(error))
         let st_ref = await firebase.storage()
           .ref(`artisanFiles/${uid}/images/profilePicture`)
           .putFile(profilePicturePath)
-
+        
         artisanObject.profilePictureURL = st_ref.downloadURL
         await firebase.database().ref(`/artisans/${uid}`).update({ profilePictureURL: st_ref.downloadURL })
       }
